ai_image_service/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
Pillow>=10.1.0
numpy==1.26.2
opencv-python-headless>=4.8.1.78
rembg[gpu]>=2.0.67
torch>=2.2.0
torchvision>=0.17.0
onnxruntime-gpu>=1.17.0
basicsr>=1.4.2
facexlib>=0.3.0
realesrgan>=0.3.0